+++ Documents/ENG D STUFF/VFX/RAY TRACER/search_tree.hpp
@@ -12,7 +12,7 @@ class search_tree{
         float parameters [6];
         int number_of_node_faces;
         static void build_tree(float* vertices, int* faces, std::vector<search_tree*> leaf_nodes, search_tree**root );
-        static void traverse_tree(search_tree*root, vector3 eye, vector3 d, std::vector<float> *output);
+        static void traverse_tree(search_tree*root, vector3 eye, vector3 d, std::vector<int> *output);
         static void leaf_nodes(float* vertices, int*faces, int number_of_faces, std::vector<search_tree*> *leaf_nodes);
         static void find_parameters(int i, float* vertices, int*faces,  std::vector<float> *parameters, std::vector<float> initial_parameters);
     private:
